Revert "Update routes to final form (#2374)"
This reverts commit e9d5463eb6ea1ae2385ea0fb047723cfca7a7545.
