+++ Classes/Systems/AppDelegate.swift
@@ -23,6 +23,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
 
         appController.appDidFinishLaunching(with: window)
+        appController.setupRoutes()
 
         // setup fabric
         Fabric.with([Crashlytics.self])
@@ -49,7 +50,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func application(_ application: UIApplication, performActionFor shortcutItem: UIApplicationShortcutItem, completionHandler: @escaping (Bool) -> Void) {
-        appController.router.handle(path: shortcutItem.type, params: shortcutItem.params)
+        appController.handle(path: shortcutItem.type, params: shortcutItem.params)
     }
 
     func applicationDidBecomeActive(_ application: UIApplication) {
