+++ Classes/Systems/AppRouter/AppController+SetupRoutes.swift
@@ -1,0 +1,22 @@
+//
+//  AppController+SetupRoutes.swift
+//  Freetime
+//
+//  Created by Ryan Nystrom on 10/7/18.
+//  Copyright © 2018 Ryan Nystrom. All rights reserved.
+//
+
+import Foundation
+import GitHawkRoutes
+
+extension AppController {
+
+    func setupRoutes() {
+        register(route: BookmarkShortcutRoute.self)
+        register(route: SwitchAccountShortcutRoute.self)
+        register(route: SearchShortcutRoute.self)
+        register(route: IssueRoute.self)
+        register(route: RepoRoute.self)
+    }
+
+}
