+++ Settings.bundle/com.mono0926.LicensePlist.plist
@@ -120,7 +120,7 @@
 			<key>File</key>
 			<string>com.mono0926.LicensePlist/GitHawkRoutes</string>
 			<key>Title</key>
-			<string>GitHawkRoutes (0.1.1)</string>
+			<string>GitHawkRoutes (0.1.0)</string>
 			<key>Type</key>
 			<string>PSChildPaneSpecifier</string>
 		</dict>
